+++ tests/Unit/Router/RouterMatchTest.php
@@ -26,18 +26,6 @@ final class RouterMatchTest extends TestCase
         });
     }
 
-    public function test_not_respond_when_the_uri_does_not_matches(): void
-    {
-        $_SERVER['REQUEST_URI'] = 'https://example.org/unexpected/uri';
-        $_SERVER['REQUEST_METHOD'] = 'GET';
-
-        $this->expectOutputString('');
-
-        Router::configure(static function (Routes $routes): void {
-            $routes->get('other/uri', FakeController::class, 'basicAction');
-        });
-    }
-
     public function test_not_respond_when_the_method_does_not_matches(): void
     {
         $_SERVER['REQUEST_URI'] = 'https://example.org/expected/uri';
